Move firebase config to the 'lib/firebase.js'
